tests/wpt/tests/websockets/constructor/007.html

<!doctype html>
<title>WebSockets: new WebSocket(url, null char)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
test(function() {
  assert_throws_dom("SyntaxError", function() {
    new WebSocket(SCHEME_DOMAIN_PORT + '/empty-message',
                  'a' + String.fromCharCode(0) + 'b')
  })
});
</script>
